Evaluate HOSTNAME on container, not on Docker Host
Using $$HOSTNAME causes docker-compose to set HOSTNAME_COMMAND: "echo $HOSTNAME" on the container, so that the HOSTNAME will be evaluated there. This results in using the container itself as the advertised host name, which is correct for this lab. The $HOSTNAME in the old file version is evaluated on the Docker Host. Some platforms muddle through with the old version, but others report LEADER_NOT_AVAILABLE errors.